requirements.txt
streamlit==1.26.0
langchain==0.0.292
openai==0.28.0
python-dotenv==1.0.0
google-api-python-client==2.125.0
duckduckgo-search==5.3.0